[SiFive][RISCV] Recognize 8-bit image composite pattern in SelectionDAG.
AggressiveInstCombine will soon start truncating
(i8 (trunc (udiv ((and (add (mul (xor (zext A), 255), (zext X)),
                 (mul (zext A), (zext Y))), 65535), 255)))

from i32 to i16 when A, X, and Y are i8. This improves the bitwidth
used for vectorization. If we don't end up vectorizing, this scalar
code will reach SelectionDAG.

Type legalization will need to widen everything to iXLen. For the udiv,
this naively requires inserting an AND with 65535. computeKnownBits
believes the input to the udiv produces a 17 bit result. The inverse
relationship of (zext A) and (xor (zext A), 255) gives a tighter upper
bound of 65025(255 * 255) which only requires 16 bits.

This patch adds a DAGCombine to remove the unneeded AND to get back
something closer to the original CodeGen before the AggresssiveInstCombine
change.

Unfortunately, there's an additional complication that this pattern appears
in a loop where (zext A) and (xor (zext A), 255) are loop invariant. We need
to sink the xor back into the loop in CodeGenPrepare so SelectionDAG can see
the pattern needed for the analysis. We don't need to sink the zext because
SelectionDAG is able to use computeKnownBits across basic blocks well enough
for this case.