[SiFive][AggressiveCombine] Narrow 8-bit image composite pattern to i16.
GB5 Camera contains this code

```
unsigned int alpha = other_row[3];
unsigned int inverse_alpha = 255 - alpha;
for (int c = 0; c < channels; c++) {
  unsigned int tmp = (other_row[c] * alpha + this_row[c] * inverse_alpha) / 255;
  this_row[c] = tmp;
}
```

Where other_row and this_row are unsiged char arrays.

C integer promotion rules dictate that the calculation of `tmp` is done
using 32-bit integers. This loop is vectorizable, but we need to use the
narrowest type we can in order to be profitable.

computeKnownBits believes `(other_row[c] * alpha + this_row[c] *
inverse_alpha)` requires 17 bits. (255 * 255) + (255*255) = 130,050.
`alpha` and `inverse_alpha` are never both 255 at the same time so the
real upper bound (255 * 255) = 65025 which fits in 16 bits.

This patch overrides the computeKnownBits result in TruncInstCombine
when this pattern is detected. The pass normally only truncates to a
legal type for the target so I had to force 16 to be used when this
pattern is detected.

Still need to improve the vectorizer cost model to get this to
vectorize. Additional changes are needed in backend to avoid a
regression if this pattern isn't vectorized.