[SiFive][RISCV] Add support for optimized vrgather.vv on sifive-x390
The x390 configurations supported in the compiler both have an optimized
vrgather.vv which builds on the optimized vrgather and vcompress
introduced in orca for x280. As a result, the way we handle scheduling
for these instructions is refactored away from using variants in favor
of using multiple scheduler models.

A future patch will split ei16 into its own sched classes, but since we
are already grouping it into the same VRGatherVV classes, we continue
with that for now.

The behavior is slightly changed on sifive-x280n for vrgather.vv for
latency because it should have been 7 + DLEN / SEW but was VL + 3.