[SiFive][Uncountable] Refactor code for find() support (#1911)
- Handle BranchInst as #88385 does
We get the uncountable exiting block from LoopVectorizationLegality. and
create the BranchOnCond recipe after all instructions are visited.
PR88385 uses `VPEarlyExitCond` and `getSpeculativeEarlyExitingBlock`.
But the name `EarlyExit` is not suitable for uncountable loops with
single exiting such as strlen(), as strlen() exits from latch and most
EarlyExit related paths in #88385 doesn't not work for strlen.
So I choose `VPDataDepExitCond` and `getCouldNotComputeExitingBlock` as
alternatives.
And when handling find(), `EarlyExitingBB` would be equal to
`CouldNotComputeExitingBB` as the
[code](https://github.com/sifive/riscv-llvm-internal/blob/e14a207e275d3d1749495fd6cee4f8224323aaa0/llvm/lib/Transforms/Vectorize/LoopVectorize.cpp#L11452)
shows.