[SiFive] Cherry-pick upstream commits for Zicfilp branches. (#1945)
This cherry-picked upstream #68292 and #89014.

#68292
When Zicfilp enabled, this avoids selecting indirect jumps to
PseudoBRIND/PseudoCALLIndirect/PseudoTAILIndirect, since they may uses
X7 as rs1 and be identified as a software guarded jump. There is an
another PR #66762 to use software guarded jump for jumptable branch.

#89014.
PseudoTail should be a software guarded branch in Ziciflp, since its
branch target is known in link time. JALR/C.JR/C.JALR with rs1 as t2 is
termed a software guarded branch. Such branches do not need to land on a
lpad instruction.