[SiFive][RISCV] Update the P400 scheduling model on VREDUCE, VISLIDE, VMV, and VRGATHER (#1947)
This is a refresh on the P400 scheduling model for the following
instructions:
  - `vred[sum|min|max].vs`
  - `vfred[usum|osum|min|max].vs`
  - `vrgather.vv`, `vcompress.vm`
  - `vmv.s.x`, `vfvmv.s.f`
  - `vslide[up|down].vx`

Based on the data measured with llvm-exegesis, these changes (when
applied on wk139 and measured against it) don't really affect SPEC2006
and SPEC2017's scores. Most the benchmarks have less than 2% score
changes except 401.bzip2 (+4.12%), 473.astar (+3.65%) and 471.omnetpp
(-2.53%). The regression on 471.omnetpp is likely to be a noise because
it doesn't even use the instructions we're tweaking here.