[SiFive][RISCV] Make RISCVMaskInstDAGMutation deterministic
SmallSet does not have a deterministic iteration order. We do not add
edges after NumVecEdges is big enough. As a result, we do not add all
possible edges. If iteration order is different, then we can add
different edge sets on different runs. This patch fixes that by using SetVector.