[SiFive][RISCV] Add a combine for (vzext_vl (sub_vl (maxu_vl X, Y), (minu_vl X, Y))) to improve sum of absolute difference. (#1969)
If this occurs as part of a sum of absolute difference pattern, we want
to push some or all of the vzext_vl to the operands of the sub by using
a vwsubu. This avoids a vzext.vf2 instruction.

This transform replaces performABSCombine that was recently removed
after being broken by upstream.