[SiFive][RISCV] The x390 Dual-VALU resource should not consume both subunits
Prior to this patch, the SiFive7NVA1OrVA2 was marked with BufferSize = 0.

Introduced in ca8eef7e3da8f750d7c7aa004fe426d1d34787ea:
> When the ProcResGroup has BufferSize=0,
>    1. if there is a subunit in the list of write resources for the
>       scheduling class, do not attempt to schedule the ProcResGroup.

This means that when the SiFive7NVA1OrVA2 is used, VA1 *and* VA2 are used.
The intention is to have VA1 *or* VA2 are used. To solve this, we leave
BufferSize as the default value of -1.

This leads to a significant speedup in https://sifive.atlassian.net/browse/ALGOR-702.

I tried to commit the test case in this patch, and it turns out that
there is no diff to it as a result of this patch. I think this is
because llvm-mca does not interpret ProcResGroup in the same way. I am
keeping the test here because we should have llvm-mca coverage of this
behavior.