[SiFive][RISCV] Model F->V pipeline latency (#2040)
We do not model vector instructions going through the scalar part of the
pipeline. To accommodate for this, we add negative read advance on the
FP operands of vector operations. We choose `-3` since the FP pipeline
is 3 cycles deep as modeled.

FP operands to vector instructions are not fully bypassed. Even though
theoretically we don't need the operand until the WB stage, when it's
sent to the vector command queue, in this design the FP operand needs to
be ready when the vector instruction is in the AG stage

This leads to a 10% improvement in
https://sifive.atlassian.net/browse/ALGOR-702 on top of #2039.

This does not have a significant impact on spec2006.

I tried adding a llvm-mca test for this, but I don't think llvm-mca
tests update latency based on ReadAdvances.