[SiFive][LoopConcat] Implementation of the Canonicalizer
This change adds feature complete details to Loop Concatenation
as well as adds the Canonicalizer which transforms local array
storage and its uses to share temp arrays to facilitate
Loop Concatenation.  It also extends enabling of the Leaf
inline threshold to facilitate removal of the general
inline threshold multiplier in some special cases specific
to LC to extend inlining of leaf functions.