[SiFive][RISCV] Replace VAADD.*VL nodes with the AVGCEIL/AVGFLOOR nodes from upstream. NFC (#2100)
Upstream is using separate nodes for the 4 combinations of
signed/unsigned rdn/rup. Tablegen fills in the rounding mode operand.

Our downstream used 2 nodes with a rounding mode operand provided during
lowering.

This patch removes our downstream nodes in favor of the upstream to
reduce the number of isel patterns in the table.

Since we also had VASUB nodes with rounding modes created by the same
code, I've adapted them to be more similar to the AVGCEIL/FLOOR nodes.
This means dropping the rounding mode operand and having tablegen fill
it in. We only need 2 nodes because we only use RDN for these. I've
renamed them to match the Neon instruction name from Recode since I
didn't have a better name that conveyed the rounding direction.