[SiFive][RISCV] Improve setcc logic in deepsjeng_r eval function (#2090)
I saw the pattern `and (seteq A, B), (seteq C, D)` in deepsjeng. We were
not optimizing it in `foldLogicOfSetCCs` because one of the `seteq`
instructions `T` was used more than once in `select Cond, T, (and T,
X)`.

This was happening on the p470/p670 since the
`!hasCondiditionalMoveFusion` check in `combineSelectAndUse`
 led us to do the combine. By skipping the combine for an i1
 select, we avoid the `select Cond, T, (and T, X)`.

Instead, we see the pattern `and (and a, (setcc b, c)), (setcc d e)`.
This pattern can be transformed to `and a, (and setcc b, c), (setcc d,
e)`. Then from this form, `foldLogicOfSetCCs` can do its
optimizations as desired.

This is an alternative to #2084