[SiFive][LV] Support FindLastIV reduction for non-constant start values of the induction variable (2/2)  (#1996)
Following #1941, introduce a new member, `SentinelValue`, in class
`RecurrenceDescriptor` to enable the use of a maximum value as a
sentinel value. This patch attempts to use the minimum value as a
sentinel. If the minimum value cannot be used as sentinel, it then tries
the maximum value as sentinel.

SCT-2979

---------

Co-authored-by: Alexey Bataev <alexey.bataev@sifive.com>