[SiFive][Reassociate] Add an option to run reassociate twice (#2208)
In which the first run doesn't do any CSE-related optimizations and only
does that in the second run. The idea is to account for the changes done by
the canonicalizations from the first run in the histogram used for
CSE-aware reassociate in the second run.

This increases the compilation time by ~5% in the worst case in SPEC2006.
Therefore, it is disabled by default.

SCT-3939.