[SiFive][RISCV] Prevent shrinking MO_GOT_GPREL_LO loads created for compact code model. (#2214)
The linker can rewrite this load into an ADD without checking the
opcode. If we were depending on the sign/zero extending behavior of the
narrower load, this ADD will not have the same semantics.

Avoid this by preventing the load from being changed.

The RISCVMCCodeEmitter.cpp are cleanups from an aborted attempt to check
the opcode when we emit the relocation. Apparently the assembler allows
the relocation on any load instruction so I couldn't do that.

Fixes SCT-4118