[SiFive][LV] Check mask != 0 before executing masked instructions (#2181)
The PR introduces simple control flow in VPlan and vectorized loop. It
implements:
- `VPIfBlock` - region-like SEME VPBlock that is similar to if-like
statement
- `VPBranchOnMask` - extended to support any other `VPBasicBlock` as
well as unconditional branches
- `optimizeConditionalRecipes` - VPlan xform that construct `VPIfBlock`
for each masked store and moves computation tree of a stored value when
possible into the constructed `VPIfBlock`. The optimization works only
iff `TTI::hasFlattenControlFlowPenalty` is `true` for the target.
Currently this function returns `true` only for P470, but can be
overridden with `riscv-has-vector-flatten-control-flow-penalty` clopt