[SiFive][LegalizeVectorTypes] Preserve original MemoryOperand and MemVT when widening fixed vector load to vp_load. (#109473) (#2246)
SIFIVE cherry-pick from 93baa018e09bb3d4d5f4da0232321aff204caaeb

Previously we were building a new memoperand with the size of the
widened VT. This was causing a failure in our downstream with non-power
of 2 vectorization. Alias analysis allowed rescheduling a 3 element
vector load past 2 out of 3 scalar stores that overwrite what it was
supposed to read.

Alias analysis considers it undefined behavior to read more than the
size of the underlying object. There is an exception if the underying
objects is sufficiently aligned, but that doesn't apply in my failing
case.

Fixes SCT-4122 and SCT-4175