[SiFive][ValuePressure][Inlining] Add support for enabling value pressure based inlining
This change enables value pressure guarding in the inliner.  During
InlineCostCallAnalyzer::finalizeAnalysis we check first the callee
value pressure to see if the initial inline candidate is already
over the TTI limit for the machine RC's, while doing so we return
the number of each class's local maxima to use later.  If the
callee is under the limits we proceed to the caller to emulate
the effect of the callee as inlinined by using the data we collected
to reduce the TTI limits by what the callee used.  We only need evaluate
the call site block of the current edge being considered.  If
over the prescribed limits, in either case, we do not inline the
candidate.  This analysis is done just before applying the final
state of the inline candidate so as to reduce compile time cycles.