[RISCV] Add an option to emit vlse from constant pool instead vmv.v.x/i (#2300)
The algorithms teams prefers this pattern for splatting constants on
x280. It uses the load sequencer instead of the arithmetic sequencer
which is usually busier in their workloads.

Unfortunately, to not compromise selection .vi and .vx arithmetic
instructions I had to the constant pool creation during isel. This
requires duplicating all of the different code model lowerings for
constant pools.

We may also want to disable formation of .vx instructions or prefer a
vlse over a scalar load from constant pool and an arithmetic .vx
instruction, but I haven't done anything for that in this patch.

This is currently off by default behind an option to enable
experimenting.

This is a step towards addressing SCT-4168.