[SiFive][RISCV] Throttle load clustering by the number of SUnit successors (#2324)
The regression in SCT-4208 was caused by load clustering in a way that
the (load) clusters were "pushed" to the beginning of the basic block,
leaving the rest of the block with anything but floating point
operations, which saturated the FEX pipe. The force that pushed those
clusters up are the newly inserted dependency edges induced by each
successor of the clustered SUnits.

Thus, this patch fixes this issue by throttling the clustering by the
number of SUnit successors. The threshold is proportional to the size of
the enclosing block. Empirical experiments found that 8% is probably a
better threshold for P470 on SPEC2006.

I implemented this feature in a more radical way (i.e. updating the
`shouldClusterMemOps` interface) because I think it could be a good
candidate for upstreaming in the future.