[SiFive][RISCV] Fix DAG mutation addition for SiFive7. (#2357)
The PreRA mutation was adding 2 store cluster mutations. The PostRA
mutation was completely overwriting the DAG created for load/store
clustering.

I've moved the SiFive7 code up and used an early return to bypass the
upstream code. This makes the code consistent with the behavior of
RISCVInstrInfo::shouldClusterMemOps which only handles stores for
SiFive7. I don't know how our downstream vector mask mutation and
upstream's createRISCVVectorMaskDAGMutation interact so it seemed best
to bypass that too.