[SiFive][RISCV] Allow 7 for VXRM operand in MachineVerifier. (#2363)
This is needed for legacy C intrinsics without VXRM argument.

Restore the few tests that we had for this that were modified in the
last pulldown. These tests weren't trying to test this feature, but they
happened to use 7.

Add a new verifier check for the VXRM implicit use that is necessary
when the value is 7.