[SiFive][RISCV] Add upstream RISCVVLOptimizer to our repo (#2338)
fd1925f1c497882b06f367befd4ac9bbb4203eec and
dc03fcd33f7f705dbcf95ec90582972f0f9ea14e reverted the addition of the
upstream RISCVVLOptimizer because it wasn't ready to be enabled
downstream. This patch adds it back in a disabled state so that we can
continue to pull down the improvements made to the upstream
RISCVVLOptimizer. This will simplify staging the switch from the
downstream to the pulled-down upstream version of the RISCVVLOptimizer.