[SiFive][SelectionDAG] Teach computeKnownBits about (mul (mul (zext (i8 X), (zext (i8 Y)), 32897). (#2375)
32897 is the magic constant emitted for (udiv X, 255) when X is
known to have 16 active bits. The udiv will be expanded to (x * 32897)
>> 23.

Naively we believe this result requires 9 bits. This is because
substituting the maximum 16-bit value 65535 for x gives
(65535 * 32897) >> 23  which simplies to 257.

If we know that x is a multiply of 2 8-bit numbers, we can put a
tighter bound of 65025 on the 16-bit value. (65025 * 32897) >> 32 is 255
which requires 8 bits instead of 9.

By teaching computeKnownBits this fact, we can use it to simplify
other code that is also looking for 8-bit values.

I believe this will improve SCT-4306.