[SiFive][RISCV] Fix ODR violation between the two different VLOptimizers. (#2379)
OperandInfo is in both files with different implementations. We need to
put it in an anonymous namespace to prevent it from being exported.

My suspicion is that all the functions are inline, but on a debug build
they won't be inlined and will get weak linkage. The linker will
silently pick one. Since the implementations are different this will
result in weird behavior if the one from the the upstream file is
picked.