[SiFive][RISCV] Remove SiFive_SiFiveRISCVVLOptimizer.cpp
Now that the upstream VLOptimizer is more mature, we remove the
SiFive_SiFiveRISCVVLOptimizer.cpp in favor of the upstream version. The
upstream version seems to be worse at optimizing when there is a tail
agnostic policy since Philip does not believe that those are safe to
optimize. We also seem to have a bug in our downstream version when
comparing a register to a constant.