[SiFive][Merge] Cherry-pick 404af37
[VPlan] Remove stale assertion in HCFG builder.

The assertion was left over from a time when VPBBs still had an
associated condition bit. This is not the case any more (comment was
stale). In case a branch on condition is needed, a BranchOnCond
VPInstruction is added when constructing recipes. That's also where it
is checked if the condition is available.

Exposed by 38376dee9.