[SiFive] Cherry-pick "[RISCV] [MachineOutliner] Analyze all candidates (#127659)" (#2409)
SIFIVE cherry-pick 6757cf4e6f1c7767d605e579930a24758c0778dc. Fixes
SCT-4361.

#117700 made a change from analyzing all the candidates to analyzing
just the first candidate before deciding to either delete or keep all of
them.

Even though the candidates all have the same instructions, the basic
blocks in which they are present are different and we will need to check
each of them before deciding whether to keep or erase them.
Particularly, `isAvailableAcrossAndOutOfSeq` checks to see if the
register (x5 in this case) is available from the end of the MBB to the
beginning of the candidate and not checking this for each candidate led
to incorrect candidates being outlined resulting in correctness issues
in a few downstream benchmarks.

Similarly, deleting all the candidates if the first one is not viable
will result in missed outlining opportunities.

Co-authored-by: Sudharsan Veeravalli <quic_svs@quicinc.com>