[SiFive][DFAJumpThreading] Disable early exit heuristics (#2424)
We disabled this feature a while ago. Then upstream tried to fix it in
5fb57131b744c52f74919f9487f4a9fa69f455fb. I verified that it recovered
the regression on the sifive-dev at that time. But unfortunately the
regression came back again after this pulldown, bringing ~2.5%
regressions to 400.perlbench on P470.

This patch disables it again. I confirmed it recovered the said
regression on 400.perlbench.