[SiFive][VPlanCM] Improve register pressure model.  (#2426)
This patch improve register pressure model in VPlanCM by following
changes to prevent register spill. 

* Calculate register pressure for widen-induction and widen-GEP recipes.

* Calculate live-in register for the VPWidenPointerInductionRecipe which
will use a live-in value for the step-offset.

* Calculate the register pressure for the WidenStore recipes if the
stored value is constant. Since we don't have vse.vi instruction, we
need to splat the constant value to a vector  and store it to the
destination.

Currently this changes is hard to implement in the upstream since
upstream is hard to know if a live-in is scalar or vector and hard to
identify the WidenPointerInduction from scalar instructions.

[SCT-4370](https://sifive.atlassian.net/browse/SCT-4370)