[SiFive][RISCV] Replace Mammoth with XSfmm in anything used by the assembler. NFC (#2441)
We need to share code with a customer and that should be in a form that
is consistent with what we would post upstream. Since Mammoth is a code
name, we should replace it with something more closely related to the
extension names.

I didn't update all references to Mammoth. Just what is needed for the
first patch to the customer.