[SiFive][Merge] Revert commit 417810a to fix VPlan verifier
In commit 417810a [SiFive][SCT-2801][LV] Generate predicate in a proper
VPBasicBlock (#1328), we move the insertion point to the end of VPBB,
causing new recipes to be added after the terminator and it is
detected by the newly added VPlan verifier.

This patch reverts the commit,
and LoopVectorize/RISCV/sifive_sct2801.ll still passes.