backport 132895 (#2473)
[RISCV] Make RequiredExtensions for intrinsics scalable to more than 32
extensions. NFC (#132895)

We have more than 32 extensions in our downstream and had to change this
type from uint32_t to uint64_t.

To simplify our downstream and make the code more flexible, I propose to
make it an array of uint32_t that we can size based on the number of
extensions. I really wanted to use std::bitset, but we have to print the
bits to a .inc file which can't easily be done with std::bitset.

Co-authored-by: Craig Topper <craig.topper@sifive.com>