[SiFive][VPlan] Align upstream vplan-based cost of Monotonic recipes to downstream vplan-based model. (#2478)
This patch add more uniform VPRecipes that can be the user of the
monotonicPHIRecipe and MonotonicUpdateRecipe, which should return a
valid cost.

https://github.com/sifive/riscv-llvm-internal/blob/3bfcae07563130623ab80c653011cdab5d5a4569/llvm/lib/Transforms/Vectorize/SiFive_VPlanCostModel.cpp#L527-L543
This change align the behavior of the upstream vplan-based cost of
VPMonotonicPHIRecipes
to the downstream.


This change is founded by SCT-4424, which shows the 445.gobmk
engine/dragon.c:1729 won't be vectorized using upstream cost model.