[SiFive][RISCV] Drop DLen related tune feature
We defined series of DLen tune features in our downstream, however we have added
DLenFactor on upstream, so we can drop the DLen related tune features
and use DLenFactor instead.

But the story is a bit more complicated, because we have few places
check for the DLen values, it can be calculate from DLenFactor in theory,
however it was only able to get the DLen when you specify the SiFive cpu or
expliclty specify the DLen tune feature, so I tried to keep the same behavior as
before with little changes on the tests.

Anyway, the main motivation of this change is reduce the difference
between upstream and downstream.