[SiFive][RISCV] Convert (mul (zext X), (splat Y)) to (mul (zext X), (zext (splat (trunc Y)))). (#2507)
If Y already has zeroes where the zext produces zeroes. This encourages
the formation of widening multiply.

We have to do this in CodeGenPrepare because the splat may be in another
basic block which will prevent SelectionDAG from seeing it.

We already do this for vp_mul/vp_zext.

SCT-4446