[SiFive][LoopUnroll] Enable runtime unroll for multi-exit loops on in-order cores (#2531)
The default unroll strategy does not fully unroll loops with multiple
exits, unless we enable 'LoopUnrollPreference::RuntimeUnrollMultiExit`,
in which case it would eliminate boundary checks (against induction
variable) in each iteration and use prolog/epilog to treat the
remainders.

This patch enables `RuntimeUnrollMultiExit ` for in-order RISC-V cores,
which not only recovers the dynamic IC regression on 483.xalancbmk, but
also yields an _additional_ 4.9% dynamic IC improvement compared to
baseline (wk191). 

SWQA-2702