[SiFive][Merge] Fixup VPConditionalRegionBlock.
In this pull-down, BranchOnMask::execute() will the the BlockInMask from
the first operand instead of `getMask()`. So we need to set the all true
mask as the first operand of the VPConditionalRegionBlock.

Also in this pull-down, the unreachable terminator will be fixed during
execution. So need to insert the created branch by builder instead of
new a branch instrution directly.

This patch is a quick fix of the VPConditionalRegionBlock. We probably
need to refactor the design of the VPConditionalRegionBlock.