[SiFive][RISCV] clean mammoth vsetvl insertion (#2529)
The idea is to separate the TM and TK registers from the VSETVLInfo
struct and only use the TWiden to determine whether the current status
is for mammoth.

1. The TN operand will be fully treated as AVL, so we can reuse all
optimizations.
2. SETTM/SETTK's vtype will also be extracted and modeled as AVLMAX()
VSETVLInfo in the original flow to preserve the vtype value.
3. SETTM and SETTK are emitted in different flows without any
optimization. I will add optimization to them when I figure out a good
way to reuse the AVL optimization for TM and TK.