[SiFive][InstCombine][SimplifyCFG] Prevent separating extractvalue from deinterleave intrinsics through phi nodes. (#2569)
The InterleavedAccessPass is dependent on the presence of the
extractvalue consuming the result of the deinterleave intrinsic. If
SimplifyCFG or InstCombine pull the extractvalue through a phi, it is no
longer a direct user so the InterleavedAccess pass won't kick in.

This can cause crashes if we're deinterleaving a vp.strided load with an
element type like i128 that isn't natively supported. The pattern we use
for a strided segment load is a vp.strided.load using the segment size
as the element type, followed by a bitcast to break up the elements in
the segment, followed by a deinterleave intrinsic.

The better fix is to harden the InterleavedAccess pass to not require
extractvalues, but instead update them if they exist. For phi uses we
can recreate the expected struct output the original deinterleave
intrinsic should have created.

Fixes SCT-4513.