[SiFive][InlineCost] Adjust the inline cost and threshold based on loops (#2565)
For every loops in the callee function, we discount the cost of the
enclosing instructions based on the loop depth; for a candidate callsite
enclosed in a loop, we increase the threshold, also based on the loop
depth.

Empirical experiments show that it's better to avoid inlining callee
with nested loops and honor callsites placed inside nested loops.

SCT-3749