[SiFive][RISCV] Add isel pattern for vmerge_vl+fma and an optional fneg. (#2523)
The default fma patterns always pick a form that clobbers one of the
multiplicands. If the addend is shared with a vmerge_vl that uses the
fma we should use the form that clobbers the accumulator instead. We
already to this for the fma_vl patterns created for VP intrinsics. This
handles the non-VP case for the fma.

Ideally the RISCVVectorPeephole pass would do this.

Fixes SCT-4476 and SCT-4533