[SiFive][RISCV] Emit pseudoinstructions for compact code model. Expand in RISCVExpandPseudoInsts. (#2603)
The %got_gprel_lo specifier should only appear on a pointer sized load. It
cannot appear on an addi feeding a separate load. The linker assumes
this.

Previously we emitted separate RISCVISD::ADD_LO and ISD::LOAD, relying on
selectAddrRegImm to fold them. VSW encountered a case where the load was
turned into a x0-strided vlse and the RISCVISD::ADD_LO was left alone
and got selected as an ADDI. The linker rewrote the ADDI thinking it was
a GOT load, causing the vlse to read from the address of the symbol
instead of the GOT.

To make this more robust, this patch directly emits all compact code
model sequences to their assembler pseudoinstructions. These will be
expanded by RISCVExpandPseudoInsts like we do for other load address
pseudoinstructions from other code models. This guarantees the load
instruction will always be created with the %got_gprel_lo specifier.

Fixes SCT-4547.