[SiFive][Merge] Fixup get SCEV step value in `fixupIVUsers()` after `c482b8faeace855`.
After c482b8faeace855 , the SCEV recipes will be executed at the initial stage of the
VPlan execution. When LV tried to fixup the IV users (in the late VPlan
execution stage) the SCEV recipes in the preheader already be replaced
and removed. So the Step value will be null or incorrect.

This patch  add a temporary map `ExpandedSCEVs` to the VPlan that cache the generated
valud of the SCEVs recipes in the preheader. Then in the
`fixupIVUsers()`, can get the genreated Value from it.

Note that not all SCEV step recipes are in the preheader, some of them
are live-in value which will not be executed. Need to get the underlying
IR value for these non-cache (live-in) SCEV recipes.