[SiFive][LV] Clean up the extraction of live-out value of fixed-order recurrence phi under tail folding with EVL. (#2589)
When performing tail folding, the active lane count in the final
iteration may be as low as one (i.e., EVL = 1). In such cases, it
becomes impossible to extract the penultimate element from the backedge
vector of a recurrence phi with only one active lane. This patch changes
the strategy to directly extract the last element from the spliced
vector of the recurrence phi and its backedge, ensuring correctness even
when EVL is 1.

Related upstream patch https://github.com/llvm/llvm-project/pull/142772