[SiFive][Merge] Remove custom opt for cond branch with cond == True for Term of middle block.
This patch remove the custom implemention of removing the terminator of
the middle block by unconditional branch when the conditional branch
with always true condition.

This optimization was introduced in upstream in `5fbd0658a02bd91cf3641`
by Florian. So remove the custom implementations.

Note that most of the test cases were not updated for this commit.