[SiFive][Merge] Prevent adding exit block into loop.
After 283a78a088bc669f31d, adding generated BB to the Loop is moved from
VPBasicBlock::execute() to connectToPredecessors(). Which means that
some of the VPIRBasicBlock will be add back to the loop. And in this
case, the BB is the exit of the original loop, if add it back, the
verifier will fail.