[SiFive][RISCV] Rearrange (shufflevector (fpext X)) if it will help form Zvfbfwma instructions. (#2631)
If we have an FMA were one multiplicand is an fp_extend of a bf16 and
the other
multiplicand is a splat shuffle of a a fp_extend from bf16, rewrite to
shuffle
the input of the fp_extend first. This will make both multiplicands of
the FMA be fp_extend as needed by Zvfbfwma.
    
First step to fix SCT-4550.