[SiFive][Merge] Fix instruction generated by SCEV expander crash at function verifier.
In the downstream customization, the VPWidenMemory recipes supports
stride which is expressed by SCEV and will be expanded when execution.
During the execution, the SCEV exapnder using the old SE with old DT
which has no information of the new created vector BBs such as vector
preheader and vector body.

The SCEV expander will try to reuse the
cached expanded value if possible (operand dominate use ...).
In this case the new created BBs(vector preheader) are not in the DT so
`dominates(<any_value>, <new_BB>)` will always return true.

To prevent this error, need to create the Scalar evolution with the new
DT after vector preheader created. Therefore the vector.preheader is not
unrecheable in the new DT, which make the domination analysis
correctly.

Note that this fix only works when the SCEV expand can be hoist to the
vector preheader. If the insertPoint of the SCEV is in the loop body,
this fix might not work since LV update the DT at the end of the
vplan execution.