[SiFive][VPlan] Remove deprecated InnerLoopVectorizer::fixupIVUsers() (#2634)
Deprecated InnerLoopVectorizer::fixupIVUsers() is cherry-picked in
cb2211936f2cce14ec8e4d5f9fff8172fe1ee8f1 because the IV liveout is not
optimized in upstream at the time. Remove
InnerLoopVectorizer::fixupIVUsers() for now since in this pulldown,
commit 5fbd0658a0 is landed and getMiddleBlock() cannot be used after
loop region is removed.

This PR is to make the changes more explicitly which is applied in
pulldown #2621.