[SiFive][VPlan] Split the stride of VPWidenMemoryRecipe as VPExpandSCEVRecipe. (#2643)
In the downstream customization, the widen-memory recipes contains
stride represent by SCEV. The SCEV value will be expanded at vector body
during plan execution. However, the DT will be updated at the end of the
plan execution, so expand the SCEV in the VPBB is dangerous.

Upstream provide a function to create VPExpandSCEV recipe and insert the
recipe to the loop entry, which already in the DT and must dominate
vector block.

This patch split the SCEV stride out of the VPWidenMemoryRecipe by
creating a new VPExpandSCEVRecipe and insert it to the loop entry. The
SCEV expansion now is handle by the SCEV recipe itself. This can help
our downstream implementation more robust and close to upstream.

Expand SCEV in the vector body cause a bug in this pull-down (#2621). I
think would be better to fix this separately.