[SiFive][VPlan][Uncountable] Separate early-exit loop from unbound loops (#2699)
This patch refines the uncountable loop check to be more restrictive,
limiting it to uncountable and unbounded loops only.

Because uncountable loops with countable latch-exiting blocks can share
the same vectorized loop skeleton and uniform analysis with early exit
loops.
And the exiting condition in countable latch-exiting blocks is now
uniform, which enables optimizations for pointer induction variables.