[SiFive][Merge] Remove vputils::isInLoopRegion().
`vputils::isInLoopRegion()` is broken after loop dissolution. And this
function is used to select initialEVL or State.EVL when generate vp.icmp
or vp.fcmp.

The check is added by 8ce9835c59bc6936d (part 1) which will generate the `vp.icmp`
in the middle block. But after 66bdc3e42ed74 (part 2), the widen
`vp.icmp` is optimized to scalar icmp. I think it's fine to remove this
check.