[SiFive][Merge][Uncountable] Separate entry and vector preheader
In this pulldown, the plan entry is expected to have a conditional
branch to vector preheader and scalar preheader.
And later the vecotr preheader VP basic block would be replaced by
the IR basic block.
Without the conditional branch, connectToPredecessors would
incorrectly place a direct jump to scalar preheader in the entry,
leaving the vector preheader unreachable.